--- UOLandscape/UI/Components/IDebugWindow.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace UOLandscape.UI.Components
-{
-    internal interface IDebugWindow: IUIComponent
-    {
-    }
-}
