+++ UOLandscape/UI/Windows/AboutWindow.cs
@@ -1,7 +1,7 @@
 ﻿using System.Reflection;
 using ImGuiNET;
 
-namespace UOLandscape.UI.Components
+namespace UOLandscape.UI.Windows
 {
     internal sealed class AboutWindow : IAboutWindow
     {
