+++ UOLandscape/UI/Windows/DebugWindow.cs
@@ -1,7 +1,7 @@
-﻿using ImGuiNET;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
+using ImGuiNET;
 
-namespace UOLandscape.UI.Components
+namespace UOLandscape.UI.Windows
 {
     internal sealed class DebugWindow : IDebugWindow
     {
