+++ UOLandscape/UI/Windows/DockSpaceWindow.cs
@@ -1,7 +1,7 @@
 ﻿using ImGuiNET;
 using Num = System.Numerics;
 
-namespace UOLandscape.UI.Components
+namespace UOLandscape.UI.Windows
 {
     internal sealed class DockSpaceWindow : IDockSpaceWindow
     {
