+++ UOLandscape/UI/Windows/IDebugWindow.cs
@@ -1,0 +1,6 @@
+﻿namespace UOLandscape.UI.Windows
+{
+    internal interface IDebugWindow: IUIComponent
+    {
+    }
+}
