+++ UOLandscape/UI/Windows/INewProjectWindow.cs
@@ -1,4 +1,4 @@
-﻿namespace UOLandscape.UI
+﻿namespace UOLandscape.UI.Windows
 {
     internal interface INewProjectWindow : IUIComponent
     {
